Refactor error-handling in parser/xml
Replace everything `error-chain` related in the XML-parser.
In addition xml.rs was refactored into smaller mods (the original file
over 1800 lines).
Furthermore, the parser-error variants of `error::Error` where removed.
