+++ sophia/src/parser/xml.rs
@@ -15,10 +15,9 @@ use quick_xml::events::BytesStart;
 use quick_xml::events::BytesText;
 use quick_xml::events::Event;
 use quick_xml::Reader;
-use quick_xml::Result as QuickXmlResult;
 use url::Url;
 
-use crate::error::*;
+// use crate::error::*;
 use crate::ns::rdf;
 use crate::ns::xsd;
 use crate::ns::Namespace;
@@ -31,6 +30,13 @@ use crate::term::matcher::TermMatcher;
 use crate::term::StaticTerm;
 use crate::term::Term;
 use crate::term::TermError;
+
+mod _error;
+pub use self::_error::*;
+mod _scope;
+pub use self::_scope::*;
+mod _handler;
+use self::_handler::*;
 
 const DEFAULT_BUFFER_SIZE: usize = 8 * 1024;
 
@@ -81,106 +87,6 @@ mod xmlname {
     }
 }
 
-// ---
-pub mod error {
-    error_chain! {
-        types {
-            Error, ErrorKind, ResultExt;
-        }
-        errors {
-            XmlError(e: ::quick_xml::Error) {
-                description("xml parser failed")
-                display("xml parser failed: {:?}", e)
-            }
-            DuplicateId(id: String) {
-                description("duplicate ID")
-                display("duplicate ID: {:?}", id)
-            }
-            UnknownNamespace(ns: String) {
-                description("unknown namespace")
-                display("unknown or undeclared namespace: {:?}", ns)
-            }
-            InvalidNodeName(n: String) {
-                description("invalid property name")
-                display("invalid property name: {:?}", n)
-            }
-            InvalidPropertyName(n: String) {
-                description("invalid property name")
-                display("invalid property name: {:?}", n)
-            }
-            InvalidAttribute(n: String) {
-                description("invalid attribute")
-                display("invalid attribute: {:?}", n)
-            }
-            InvalidXmlName(n: String) {
-                description("invalid XML name")
-                display("invalid XML name: {:?}", n)
-            }
-            InvalidParseType(ty: String) {
-                description("invalid `parseType` value in this context")
-                display("cannot use `parseType={}` in this context", ty)
-            }
-            AmbiguousSubject {
-                description("cannot have `rdf:ID`, `rdf:nodeID` and `rdf:about` at the same time")
-                display("cannot have `rdf:ID`, `rdf:nodeID` and `rdf:about` at the same time")
-            }
-            UnexpectedEvent(exp: String, found: String) {
-                description("unexpected XML event")
-                display("unexpected XML event: expected {}, found {}", exp, found)
-            }
-            NoBaseIri(iri: String) {
-                description("document does not define a base IRI")
-                display("base IRI needed to expand: {:?}", iri)
-            }
-            InvalidPrefix(p: String) {
-                description("invalid prefix")
-                display("{:?} cannot be used as a prefix", p)
-            }
-        }
-    }
-
-    impl From<quick_xml::Error> for Error {
-        fn from(e: quick_xml::Error) -> Self {
-            Self::from_kind(ErrorKind::XmlError(e))
-        }
-    }
-
-    impl From<quick_xml::Error> for crate::error::Error {
-        fn from(e: quick_xml::Error) -> Self {
-            Error::from(e).into()
-        }
-    }
-
-    impl From<Error> for crate::error::Error {
-        fn from(error: Error) -> Self {
-            let message = error.0.to_string();
-            crate::error::Error::with_chain(
-                error,
-                crate::error::ErrorKind::ParserError(message, crate::parser::Location::Unknown),
-            )
-        }
-    }
-
-    impl From<ErrorKind> for crate::error::Error {
-        fn from(kind: ErrorKind) -> Self {
-            let error = Error::from_kind(kind);
-            Self::from(error)
-        }
-    }
-
-    use crate::parser::Location;
-
-    /// Patch the `location` field of the error kind if the error is a `ParserError`.
-    pub fn with_position(mut e: crate::error::Error, n: usize) -> crate::error::Error {
-        if let crate::error::ErrorKind::ParserError(_, ref mut location) = e.0 {
-            *location = Location::from_offset(n);
-        }
-        e
-    }
-}
-
-use self::error::ErrorKind as XmlErrorKind;
-
 /// RDF/XML parser.
 #[derive(Clone, Debug, Default)]
 pub struct RdfXmlParser {
@@ -191,7 +97,7 @@ impl RdfXmlParser {
     pub fn with_base(base: &str) -> Result<Self> {
         match Url::parse(base) {
             Ok(url) => Ok(Self { base: Some(url) }),
-            Err(_) => Err(TermError::InvalidIri(base.to_owned()).into()),
+            Err(_) => Err(RdfError::InvalidBaseIri(base.to_owned()).into()),
         }
     }
 }
@@ -210,30 +116,6 @@ def_mod_functions_for_bufread_parser!(RdfXmlParser);
 
 // ---
 
-/// The state of the parser.
-#[derive(Debug, Clone, Copy)]
-enum ParsingState {
-    /// The parser is in a predicate, and the next expected element is a node.
-    Node,
-    /// The parser is in a node, and the next expected element is a predicate.
-    Predicate,
-    /// The parser is in a resource predicate, and the next expected element
-    /// is a predicate.
-    Resource,
-    /// The parser is in a predicate and will process its content as a literal.
-    Literal, // NB: not supported by quick-xml right now
-    /// The parser is in a reified element, and the next expected element is
-    /// a text to be used as a reified triple object.
-    Res,
-    /// The parser is in a collection, and the next expected element is a node.
-    Collection,
-    /// The parser is in a collection, and the next expected element is a node,
-    /// but exiting this does not exit the collection.
-    CollectionItem,
-}
-
-// ---
-
 /// A wrapper for `quick_xml::Reader` ignoring or merging some events.
 pub struct XmlReader<B: BufRead> {
     inner: Reader<B>,
@@ -243,7 +125,7 @@ pub struct XmlReader<B: BufRead> {
 
 impl<B: BufRead> XmlReader<B> {
     /// Read an XML event.
-    pub fn read_event<'a>(&mut self, buf: &'a mut Vec<u8>) -> QuickXmlResult<Event<'a>> {
+    pub fn read_event<'a>(&mut self, buf: &'a mut Vec<u8>) -> Result<Event<'a>, quick_xml::Error> {
         use quick_xml::events::Event::*;
 
         // Clear the event peeking cache if it is not empty.
@@ -309,249 +191,6 @@ impl<B: BufRead> Deref for XmlReader<B> {
     }
 }
 
-// ---
-
-/// Data relevant to an XML scope.
-#[derive(Debug)]
-pub struct Scope<F: TermFactory> {
-    /// The XML namespaces declared in this scope.
-    ns: HashMap<String, Namespace<F::TermData>>,
-    /// The default XML namespace to expand tags without namespaces with.
-    default: Option<Namespace<F::TermData>>,
-    /// The base IRI namespace to expand `rdf:ID`, `rdf:resource` and `rdf:about`.
-    base: Option<Url>,
-    /// The term factory used to create new terms.
-    factory: Rc<RefCell<F>>,
-    /// The datatype of the containing element.
-    datatype: Option<Term<F::TermData>>,
-    /// The language tag of the containing element.
-    lang: Option<F::TermData>,
-    /// The text gathered in the current scope.
-    text: Option<String>,
-    /// The current count of list elements
-    li: AtomicUsize,
-    /// The
-    collection: Vec<Term<F::TermData>>,
-}
-
-// We implement it ourselves instead of deriving so that:
-// * F does not need to be `Clone` (deriving requires it).
-// * we can clone `li` although `AtomicUsize` is not `Clone`.
-impl<F: TermFactory> Clone for Scope<F> {
-    fn clone(&self) -> Self {
-        Self {
-            ns: self.ns.clone(),
-            default: self.default.clone(),
-            base: self.base.clone(),
-            factory: self.factory.clone(),
-            datatype: self.datatype.clone(),
-            lang: self.lang.clone(),
-            text: self.text.clone(),
-            li: AtomicUsize::new(self.li.load(Ordering::Relaxed)),
-            collection: self.collection.clone(),
-        }
-    }
-}
-
-impl<F: TermFactory> Scope<F> {
-    /// Create a new `Scope` with the given term factory.
-    fn with_factory(f: F) -> Self {
-        Self::with_factory_rc(Rc::new(RefCell::new(f)))
-    }
-
-    /// Create a new `Scope` from a shared smartpointer to a term factory.
-    fn with_factory_rc(f: Rc<RefCell<F>>) -> Self {
-        let mut scope = Self {
-            ns: HashMap::new(),
-            default: None,
-            base: None,
-            factory: f,
-            datatype: None,
-            lang: None,
-            text: None,
-            li: AtomicUsize::new(1),
-            collection: Vec::new(),
-        };
-        // These namespaces are always in scope
-        scope
-            .add_prefix("xml", "http://www.w3.org/XML/1998/namespace#")
-            .unwrap();
-        scope
-            .add_prefix("xmlms", "https://www.w3.org/2000/xmlns/")
-            .unwrap();
-        scope
-    }
-
-    /// Add a new XML prefix to the namespace mapping.
-    fn add_prefix(&mut self, prefix: &str, value: &str) -> Result<()> {
-        if prefix == "_" {
-            bail!(XmlErrorKind::InvalidPrefix(prefix.into()))
-        } else {
-            let mut f = self.factory.borrow_mut();
-            self.ns.insert(
-                String::from(prefix),
-                Namespace::new(f.get_term_data(value))?,
-            );
-            Ok(())
-        }
-    }
-
-    /// Set the default XML prefix.
-    fn set_default(&mut self, default: &str) -> Result<()> {
-        let mut f = self.factory.borrow_mut();
-        self.default = Some(Namespace::new(f.get_term_data(default))?);
-        Ok(())
-    }
-
-    /// Set the base IRI prefix.
-    fn set_base(&mut self, base: &str) -> Result<()> {
-        // Accept the URL only if it is a valid URL and a valid base.
-        if let Ok(url) = Url::parse(base) {
-            if !url.cannot_be_a_base() {
-                self.base = Some(url);
-                return Ok(());
-            }
-        }
-
-        bail!(TermError::InvalidIri(String::from(base)))
-    }
-
-    /// Set the scope datatype.
-    fn set_datatype(&mut self, datatype: &str) -> Result<()> {
-        self.datatype = Some(self.expand_iri(datatype)?);
-        Ok(())
-    }
-
-    /// Set the scope text.
-    fn set_text<T: Into<Option<String>>>(&mut self, text: T) {
-        self.text = text.into();
-    }
-
-    /// Expand an XML attribute in the form `namespace:id` into an IRI.
-    ///
-    /// This uses the `xmlns` default namespace to expand local attributes,
-    /// or any declared namespace in the current scope.
-    fn expand_attribute(&self, attr: &str) -> Result<Term<F::TermData>> {
-        if let Some(separator_idx) = attr.chars().position(|c| c == ':') {
-            let prefix = &attr[..separator_idx];
-            let reference = &attr[separator_idx + 1..];
-            if let Some(ns) = self.ns.get(prefix) {
-                ns.get(self.factory.borrow_mut().get_term_data(reference))
-                    .map_err(Into::into)
-            } else {
-                bail!(XmlErrorKind::UnknownNamespace(prefix.to_string()))
-            }
-        } else if let Some(ns) = &self.default {
-            ns.get(self.factory.borrow_mut().get_term_data(attr))
-                .map_err(Into::into)
-        } else {
-            bail!(XmlErrorKind::UnknownNamespace("_".to_string()))
-        }
-    }
-
-    /// Expand an IRI reference (in a `rdf:resource` or `rdf:about`) into an IRI.
-    ///
-    /// This uses `xml:base` to expand local resources, and does nothing in
-    /// case the IRI is already in expanded form.
-    fn expand_iri(&self, iri: &str) -> Result<Term<F::TermData>> {
-        let mut factory = self.factory.borrow_mut();
-        if is_relative_iri_ref(iri) {
-            // NB: We should not be percent-encoding, but `url::Url::parse`
-            // does it anyway: as a fudge, we percent-decode any input that
-            // contained non-ASCII characters back. This may cause strange
-            // behaviour with URLs that contain a mix of percent-encoded and
-            // raw Unicode characters, but this is the best we can do without
-            // reimplementing the `url` crate from scratch.
-            let ascii = iri.chars().all(|c| c.is_ascii());
-
-            fn decode(s: &str) -> std::borrow::Cow<str> {
-                percent_encoding::percent_decode(s.as_bytes())
-                    .decode_utf8()
-                    .expect("always OK since validated with `is_relative_iri_ref`")
-            }
-
-            if let Some(url) = &self.base {
-                match url.join(iri) {
-                    Ok(ref u) if ascii => factory.iri(u).map_err(Into::into),
-                    Ok(ref u) => factory.iri(decode(u.as_ref())).map_err(Into::into),
-                    Err(_) => bail!(TermError::InvalidIri(String::from(iri))),
-                }
-            } else {
-                bail!(XmlErrorKind::NoBaseIri(iri.to_string()))
-            }
-        } else if is_absolute_iri_ref(iri) {
-            factory.iri(iri).map_err(Into::into)
-        } else {
-            bail!(TermError::InvalidIri(String::from(iri)))
-        }
-    }
-
-    /// Expand an ID (in a `rdf:ID`) into an IRI.
-    ///
-    /// This also uses `xml:base` to expand local resources, and prefixes
-    /// identifiers in the document with a `#` if needed.
-    fn expand_id(&self, id: &str) -> Result<Term<F::TermData>> {
-        if !xmlname::is_valid_xmlname(id) {
-            bail!(XmlErrorKind::InvalidXmlName(id.to_string()))
-        } else if id.starts_with('#') {
-            self.expand_iri(id)
-        } else {
-            self.expand_iri(&format!("#{}", id))
-        }
-    }
-
-    /// Create a new literal with the `rdf:type` and `xml:lang` in scope.
-    fn new_literal(&self, text: String) -> Result<Term<F::TermData>> {
-        match (&self.datatype, &self.lang) {
-            (Some(dt), _) => self
-                .factory
-                .borrow_mut()
-                .literal_dt(text, dt.clone())
-                .map_err(Into::into),
-            (None, Some(l)) => self
-                .factory
-                .borrow_mut()
-                .literal_lang(text, l.clone())
-                .map_err(Into::into),
-            _ => self
-                .factory
-                .borrow_mut()
-                .literal_dt(text, xsd::string)
-                .map_err(Into::into),
-        }
-    }
-
-    /// Create a new `rdf:li` property by incrementing the scope `li` counter.
-    fn new_li(&self) -> Result<Term<F::TermData>> {
-        if let Some(ns) = self.ns.get("rdf") {
-            let mut f = self.factory.borrow_mut();
-            ns.get(f.get_term_data(&format!("_{}", self.li.fetch_add(1, Ordering::Relaxed))))
-                .map_err(Into::into)
-        } else {
-            bail!(XmlErrorKind::UnknownNamespace("rdf".to_string()))
-        }
-    }
-
-    /// Get the current `rdf:li` property.
-    fn current_li(&self) -> Result<Term<F::TermData>> {
-        if let Some(ns) = self.ns.get("rdf") {
-            let mut f = self.factory.borrow_mut();
-            ns.get(f.get_term_data(&format!("_{}", self.li.load(Ordering::Relaxed) - 1)))
-                .map_err(Into::into)
-        } else {
-            bail!(XmlErrorKind::UnknownNamespace("rdf".to_string()))
-        }
-    }
-}
-
-impl<F: TermFactory + Default> Default for Scope<F> {
-    fn default() -> Self {
-        Self::with_factory(Default::default())
-    }
-}
-
-// ---
-
 /// The triple source returned by RdfXmlParser.
 pub struct RdfXmlSource<B: BufRead, F: TermFactory> {
     handler: XmlHandler<B, F>,
@@ -580,674 +219,6 @@ where
     }
 }
 
-/// An XML parser supporting any term factory as a backend.
-struct XmlHandler<B: BufRead, F: TermFactory> {
-    /// The underlying XML reader.
-    reader: XmlReader<B>,
-
-    /// The stack of scoped data (for nested declaration).
-    scopes: Vec<Scope<F>>,
-
-    /// The stack of parent elements (for nested declarations).
-    parents: Vec<Term<F::TermData>>,
-
-    // The queue of produced triples.
-    triples: LinkedList<Result<[Term<F::TermData>; 3]>>,
-
-    //
-    factory: Rc<RefCell<F>>,
-
-    //
-    bnodes: AtomicUsize,
-
-    //
-    ids: HashSet<Term<F::TermData>>,
-
-    /// The current state of the parser.
-    state: Vec<ParsingState>,
-}
-
-impl<B, F> XmlHandler<B, F>
-where
-    B: BufRead,
-    F: TermFactory + Clone + Default,
-{
-    // ---
-
-    /// Get a reference to the current scope.
-    fn scope(&self) -> &Scope<F> {
-        &self.scopes[self.scopes.len() - 1]
-    }
-
-    /// Get a mutable reference to the current scope.
-    fn scope_mut(&mut self) -> &mut Scope<F> {
-        let l = self.scopes.len();
-        &mut self.scopes[l - 1]
-    }
-
-    /// Get a reference to the parent scope.
-    fn parent_scope(&self) -> &Scope<F> {
-        &self.scopes[self.scopes.len() - 2]
-    }
-
-    /// Get a mutable reference to the current scope.
-    fn parent_scope_mut(&mut self) -> &mut Scope<F> {
-        let l = self.scopes.len();
-        &mut self.scopes[l - 2]
-    }
-
-    // Enter a new scope, parsing `xml:lang`, `xmlns` and `rdf:datatype`.
-    fn enter_scope(&mut self, e: &BytesStart) -> Result<()> {
-        // We are entering a new elements: text is not relevant anymore.
-        let mut prev = self.scope_mut();
-        prev.text = None;
-
-        // Create a local scope using values from the outer one.
-        let mut scope = prev.clone();
-        scope.text = Some(String::new());
-        scope.collection = Vec::new();
-        scope.li.store(1, Ordering::Relaxed);
-
-        // * Update XML namespaces with those defined in the document.
-        // * Change scope language if there is any `xml:lang` attribute
-        // * Fail if there is an invalid `rdf:li` attribute
-        for attr in e.attributes().with_checks(true) {
-            let a = attr?;
-            if a.key.starts_with(b"xmlns:") {
-                scope.add_prefix(
-                    &self.reader.decode(&a.key[6..]),
-                    &a.unescape_and_decode_value(&self.reader)?,
-                )?;
-            } else if a.key == b"xmlns" {
-                scope.set_default(&a.unescape_and_decode_value(&self.reader)?)?;
-            } else if a.key == b"xml:base" {
-                scope.set_base(&a.unescape_and_decode_value(&self.reader)?)?;
-            } else if a.key == b"xml:lang" {
-                scope.lang = if a.value.is_empty() {
-                    None
-                } else {
-                    let v = &a.unescape_and_decode_value(&self.reader)?;
-                    self.factory.borrow_mut().get_term_data(v).into()
-                };
-            }
-        }
-
-        // Make the newly created scope the local one.
-        self.scopes.push(scope);
-        Ok(())
-    }
-
-    // Exit the local scope.
-    fn leave_scope(&mut self) {
-        self.scopes.pop().expect("XML is not balanced");
-    }
-
-    // ---
-
-    /// Create a new bnode term (using `n` prefix).
-    fn new_bnode(&self) -> Term<F::TermData> {
-        self.factory
-            .borrow_mut()
-            .bnode(&format!("n{}", self.bnodes.fetch_add(1, Ordering::Relaxed)))
-            .expect("always produces a correct BNode")
-    }
-
-    /// Rename a bnode using the `nodeID` in the document (using `o` prefix)
-    fn rename_bnode(&self, id: &str) -> Result<Term<F::TermData>> {
-        if xmlname::is_valid_xmlname(id) {
-            self.factory
-                .borrow_mut()
-                .bnode(&format!("o{}", id))
-                .map_err(Into::into)
-        } else {
-            bail!(XmlErrorKind::InvalidXmlName(id.to_string()))
-        }
-    }
-
-    /// Check the given `ID` is unique.
-    fn check_unique_id(&mut self, id: Term<F::TermData>) -> Result<Term<F::TermData>> {
-        if self.ids.contains(&id) {
-            bail!(XmlErrorKind::DuplicateId(id.value()))
-        } else {
-            self.ids.insert(id.clone());
-            Ok(id)
-        }
-    }
-
-    /// Create a new predicate IRI from an XML name (or a RDF metasyntactic element)
-    fn predicate_iri_start(&self, name: &str) -> Result<Term<F::TermData>> {
-        let p = self.scope().expand_attribute(name)?;
-        if p.matches(&rdf::li) {
-            self.parent_scope().new_li()
-        } else {
-            Ok(p)
-        }
-    }
-
-    /// Retrieve a predicate IRI from an XML name
-    fn predicate_iri_end(&self, name: &str) -> Result<Term<F::TermData>> {
-        let p = self.scope().expand_attribute(name)?;
-        if p.matches(&rdf::li) {
-            self.parent_scope().current_li()
-        } else {
-            Ok(p)
-        }
-    }
-
-    // ---
-
-    /// Create a new `XmlHandler` from the given `quick_xml::Reader`.
-    fn new(reader: Reader<B>) -> Self {
-        let factory: Rc<RefCell<F>> = Default::default();
-        Self {
-            reader: XmlReader::from(reader),
-            parents: Vec::new(),
-            scopes: vec![Scope::with_factory_rc(factory.clone())],
-            triples: LinkedList::new(),
-            factory,
-            bnodes: AtomicUsize::new(0),
-            state: vec![ParsingState::Node],
-            ids: HashSet::new(),
-        }
-    }
-
-    /// Create a new `XmlHandler` using the given URL as the top-level `xml:base`.
-    fn with_base(reader: Reader<B>, base: Url) -> Self {
-        let mut parser = Self::new(reader);
-        let mut scope = parser.scope_mut();
-        scope.base = Some(base);
-        parser
-    }
-
-    // ---
-
-    fn element_start(&mut self, e: &BytesStart) {
-        if let Err(e) = self.enter_scope(e) {
-            let e = error::with_position(e, self.reader.buffer_position());
-            self.triples.push_back(Err(e));
-        }
-
-        if let Err(e) = match self.state.last().unwrap() {
-            ParsingState::Node => self.node_start(e),
-            ParsingState::Predicate => self.predicate_start(e),
-            ParsingState::Resource => self.predicate_start(e),
-            ParsingState::Collection => self.collection_start(e),
-            ParsingState::CollectionItem => self.collection_item_start(e),
-            ParsingState::Literal => unimplemented!("entering element as literal"),
-            ParsingState::Res => Err(Error::from(XmlErrorKind::UnexpectedEvent(
-                format!("<{}>", self.reader.decode(e.name())),
-                "text".to_string(),
-            ))),
-        } {
-            let e = error::with_position(e, self.reader.buffer_position());
-            self.triples.push_back(Err(e));
-        }
-    }
-
-    fn node_start(&mut self, e: &BytesStart) -> Result<()> {
-        // Get node type from the XML attribute.
-        let ty = self
-            .scope()
-            .expand_attribute(&self.reader.decode(e.name()))?;
-
-        // Return early if in a top-level rdf:RDF element
-        if rdf::RDF.matches(&ty) && self.parents.is_empty() {
-            self.state.push(ParsingState::Node);
-            self.parents.push(self.factory.borrow_mut().copy(&rdf::RDF));
-            return Ok(());
-        }
-
-        // Bail out if the node has an invalid name.
-        if RESERVED_NODE_NAMES.matches(&ty) {
-            bail!(XmlErrorKind::InvalidNodeName(ty.value()))
-        }
-
-        // Separate node subject from other attributes
-        let mut properties = HashMap::new();
-        let mut subject = Vec::new();
-        for attr in e.attributes().with_checks(true) {
-            let a = attr?;
-
-            // ignore xml attributes (processed in element_start)
-            if a.key.starts_with(b"xml") {
-                continue;
-            }
-
-            // try to extract the subject annotation
-            let k = self.scope().expand_attribute(&self.reader.decode(a.key))?;
-            let v = a.unescape_and_decode_value(&self.reader)?;
-
-            if k.matches(&rdf::about) {
-                subject.push(self.scope().expand_iri(&v)?);
-            } else if k.matches(&rdf::ID) {
-                let id = self.scope().expand_id(&v)?;
-                subject.push(self.check_unique_id(id)?);
-            } else if k.matches(&rdf::nodeID) {
-                subject.push(self.rename_bnode(&v)?);
-            } else if k.matches(&rdf::type_) {
-                properties.insert(k, self.scope().expand_iri(&v)?);
-            } else if RESERVED_ATTRIBUTES_NAMES.matches(&k) {
-                bail!(self::error::ErrorKind::InvalidAttribute(k.value()));
-            } else {
-                properties.insert(k, self.scope().new_literal(v)?);
-            }
-        }
-
-        // Get subject and add it to the current nested stack
-        ensure!(subject.len() < 2, XmlErrorKind::AmbiguousSubject);
-        let s: Term<_> = subject.pop().unwrap_or_else(|| self.new_bnode());
-        self.parents.push(s.clone());
-
-        // Add the type as a triple if it is not `rdf:Description`
-        if !ty.matches(&rdf::Description) {
-            let type_ = self.factory.borrow_mut().copy(&rdf::type_);
-            self.triples.push_back(Ok([s.clone(), type_, ty]));
-        }
-
-        // Add triples described by properties in XML attributes
-        for (p, lit) in properties {
-            self.triples.push_back(Ok([s.clone(), p, lit]))
-        }
-
-        // Next start event is expected to be a predicate
-        self.state.push(ParsingState::Predicate);
-        Ok(())
-    }
-
-    fn predicate_start(&mut self, e: &BytesStart) -> Result<()> {
-        // Get the predicate and add it to the current nested stack
-        // or build a new `rdf:_n` IRI if the predicate is `rdf:li`.
-        let pred = self.predicate_iri_start(&self.reader.decode(e.name()))?;
-
-        // Fail if the property is among forbidden names.
-        if RESERVED_PROPERTY_NAMES.matches(&pred) {
-            bail!(XmlErrorKind::InvalidNodeName(pred.value()))
-        } else {
-            self.parents.push(pred);
-        }
-
-        // Extract attributes relevant to the RDF syntax
-        let mut attributes = HashMap::new();
-        let mut next_state = ParsingState::Node;
-        let mut object = Vec::with_capacity(1);
-        for attr in e.attributes().with_checks(true) {
-            let a = attr?;
-
-            // Ignore `xml` attributes
-            if a.key.starts_with(b"xml") {
-                continue;
-            }
-
-            let k = self.scope().expand_attribute(&self.reader.decode(a.key))?;
-            if k.matches(&rdf::datatype) {
-                let v = a.unescape_and_decode_value(&self.reader)?;
-                self.scope_mut().set_datatype(&v)?;
-            } else if k.matches(&rdf::ID) {
-                let v = a.unescape_and_decode_value(&self.reader)?;
-                let id = self.scope().expand_id(&v)?;
-                object.push(self.check_unique_id(id)?);
-                next_state = ParsingState::Res;
-            } else if k.matches(&rdf::resource) {
-                let v = a.unescape_and_decode_value(&self.reader)?;
-                object.push(self.scope().expand_iri(&v)?);
-                next_state = ParsingState::Predicate;
-            } else if k.matches(&rdf::parseType) {
-                match a.value.as_ref() {
-                    b"Resource" => {
-                        if object.is_empty() {
-                            object.push(self.new_bnode());
-                        }
-                        self.scope_mut().set_text(None);
-                        next_state = ParsingState::Resource;
-                    }
-                    b"Collection" => {
-                        next_state = ParsingState::Collection;
-                    }
-                    b"Literal" => {
-                        self.scope_mut().set_datatype(&rdf::XMLLiteral.value())?;
-                        next_state = ParsingState::Literal;
-                    }
-                    other => {
-                        let ty = String::from_utf8_lossy(other).to_string();
-                        bail!(XmlErrorKind::InvalidParseType(ty));
-                    }
-                }
-            } else if RESERVED_ATTRIBUTES_NAMES.matches(&k) {
-                bail!(XmlErrorKind::InvalidAttribute(k.value()));
-            } else {
-                let v = a.unescape_and_decode_value(&self.reader)?;
-                attributes.insert(k, self.scope().new_literal(v)?);
-                next_state = ParsingState::Resource;
-            }
-        }
-
-        // Extract object of the triple
-        let o = match object.len() {
-            0 if !attributes.is_empty() => Some(self.new_bnode()),
-            0 if attributes.is_empty() => None,
-            1 => Some(object.last().unwrap().clone()),
-            _ => bail!(XmlErrorKind::AmbiguousSubject),
-        };
-
-        // Make the predicate a resource element if an objec tis present.
-        if let Some(o) = o {
-            self.parents.push(o.clone());
-            std::mem::replace(self.state.last_mut().unwrap(), ParsingState::Resource);
-            for (k, v) in attributes.into_iter() {
-                self.triples.push_back(Ok([o.clone(), k, v]));
-            }
-        }
-
-        self.state.push(next_state);
-        Ok(())
-    }
-
-    fn collection_start(&mut self, e: &BytesStart) -> Result<()> {
-        self.state.push(ParsingState::CollectionItem);
-        self.collection_item_start(e)
-    }
-
-    fn collection_item_start(&mut self, e: &BytesStart) -> Result<()> {
-        // Start the inner node element and get its IRI.
-        self.node_start(e)?;
-        let new_iri = self.parents.last().unwrap().clone();
-        // Add the iri of the node to the parent scope (not current!)
-        self.parent_scope_mut().collection.push(new_iri);
-        Ok(())
-    }
-
-    // ---
-
-    fn element_end(&mut self, e: &BytesEnd) {
-        if let Err(e) = match self.state.pop().unwrap() {
-            ParsingState::Node => self.predicate_end(e),
-            ParsingState::Predicate => self.node_end(),
-            ParsingState::Literal => self.predicate_end(e),
-            ParsingState::Resource => self.resource_end(e),
-            ParsingState::CollectionItem => self.collection_item_end(),
-            ParsingState::Collection => self.collection_end(e),
-            ParsingState::Res => self.res_end(),
-        } {
-            let e = error::with_position(e, self.reader.buffer_position());
-            self.triples.push_back(Err(e));
-        }
-        self.leave_scope();
-    }
-
-    fn node_end(&mut self) -> Result<()> {
-        // Add the entity as a triple object if it is not top-level
-        let o = self.parents.pop().unwrap();
-        if self.parents.len() > 2 {
-            let s = &self.parents[self.parents.len() - 2];
-            let p = &self.parents[self.parents.len() - 1];
-            if !s.matches(&rdf::RDF) {
-                self.triples.push_back(Ok([s.clone(), p.clone(), o]));
-            }
-        }
-
-        Ok(())
-    }
-
-    fn predicate_end(&mut self, e: &BytesEnd) -> Result<()> {
-        // Build the predicate IRI
-        let p = self.predicate_iri_end(&self.reader.decode(e.name()))?;
-
-        // Get the literal value
-        if self.parents.len() > 1 {
-            if let Some(text) = self.scope_mut().text.take() {
-                let s = self.parents[self.parents.len() - 2].clone();
-                let o = self.scope_mut().new_literal(text)?;
-                self.triples.push_back(Ok([s, p, o]));
-            }
-        }
-
-        self.parents.pop();
-        Ok(())
-    }
-
-    fn resource_end(&mut self, e: &BytesEnd) -> Result<()> {
-        // End of the implicit node element
-        self.node_end()?;
-        // Drop text, since it is not relevant in a Resource predicate.
-        self.scope_mut().text.take();
-        // End of the resource predicate
-        self.predicate_end(e)
-    }
-
-    fn collection_item_end(&mut self) -> Result<()> {
-        // End of the node parent.
-        self.parents.pop();
-        // Remove `CollectionItem`
-        self.state.pop();
-        Ok(())
-    }
-
-    fn collection_end(&mut self, e: &BytesEnd) -> Result<()> {
-        let collection = self.scope().collection.clone();
-        if !collection.is_empty() {
-            let mut node = self.new_bnode();
-            let mut elements = collection.into_iter().peekable();
-
-            self.triples.push_back(Ok([
-                self.parents[self.parents.len() - 2].clone(),
-                self.parents[self.parents.len() - 1].clone(),
-                node.clone(),
-            ]));
-
-            while let Some(e) = elements.next() {
-                self.triples.push_back(Ok([
-                    node.clone(),
-                    self.factory.borrow_mut().copy(&rdf::first),
-                    e.clone(),
-                ]));
-                if elements.peek().is_some() {
-                    let next_node = self.new_bnode();
-                    self.triples.push_back(Ok([
-                        node,
-                        self.factory.borrow_mut().copy(&rdf::rest),
-                        next_node.clone(),
-                    ]));
-                    node = next_node;
-                } else {
-                    let mut f = self.factory.borrow_mut();
-                    self.triples.push_back(Ok([
-                        node.clone(),
-                        f.copy(&rdf::rest),
-                        f.copy(&rdf::nil),
-                    ]));
-                }
-            }
-        }
-
-        self.predicate_end(e)
-    }
-
-    fn res_end(&mut self) -> Result<()> {
-        // Subject, predicate, object and ID of the reified triple
-        let id = self.parents.pop().unwrap();
-        let pred = self.parents.pop().unwrap();
-        let sbj = self.parents.last().unwrap().clone();
-        let txt = self.scope_mut().text.take().unwrap_or_default();
-        let obj = self.scope().new_literal(txt)?;
-
-        // Add all triples
-        self.triples
-            .push_back(Ok([sbj.clone(), pred.clone(), obj.clone()]));
-        self.reify(id, sbj, pred, obj);
-
-        Ok(())
-    }
-
-    // --- Text elements ----------------------------------------------------
-
-    fn element_text(&mut self, e: &BytesText) {
-        if self.scope().text.is_some() {
-            match e.unescape_and_decode(&self.reader) {
-                Ok(text) => self.scope_mut().set_text(text),
-                Err(e) => {
-                    let e = error::with_position(Error::from(e), self.reader.buffer_position());
-                    self.triples.push_back(Err(e))
-                }
-            }
-        }
-    }
-
-    // --- Empty elements ----------------------------------------------------
-
-    fn element_empty(&mut self, e: &BytesStart) {
-        if let Err(e) = self.enter_scope(e) {
-            let e = error::with_position(e, self.reader.buffer_position());
-            self.triples.push_back(Err(e));
-        }
-
-        if let Err(e) = match self.state.last().unwrap() {
-            ParsingState::Node => self.node_empty(e),
-            ParsingState::Predicate => self.predicate_empty(e),
-            ParsingState::Resource => self.resource_empty(e),
-            ParsingState::Collection => self.collection_item_empty(e),
-            ParsingState::CollectionItem => unreachable!(),
-            ParsingState::Literal => unimplemented!("empty element as literal"),
-            ParsingState::Res => {
-                let kind = self::error::ErrorKind::UnexpectedEvent(
-                    format!("<{}/>", self.reader.decode(e.name())),
-                    "end".to_string(),
-                );
-                Err(Error::from(kind))
-            }
-        } {
-            let e = error::with_position(e, self.reader.buffer_position());
-            self.triples.push_back(Err(e));
-        }
-
-        self.leave_scope();
-    }
-
-    fn node_empty(&mut self, e: &BytesStart) -> Result<()> {
-        self.node_start(e)?;
-        self.state.pop();
-        self.node_end()
-    }
-
-    fn predicate_empty(&mut self, e: &BytesStart) -> Result<()> {
-        let pred = self.predicate_iri_start(&self.reader.decode(e.name()))?;
-
-        // Fail if the property is among forbidden names.
-        if RESERVED_PROPERTY_NAMES.matches(&pred) {
-            bail!(XmlErrorKind::InvalidNodeName(pred.value()));
-        }
-
-        let mut object = Vec::with_capacity(1);
-        let mut attributes = HashMap::new();
-        let mut parse_type = None;
-        let mut reification = None;
-
-        // Extract attributes
-        for attr in e.attributes().with_checks(true) {
-            let a = attr?;
-
-            // ignore XML attributes (processed when entering scope)
-            if a.key.starts_with(b"xml") {
-                continue;
-            }
-
-            // try to extract the annotation object
-            let k = self.scope().expand_attribute(&self.reader.decode(a.key))?;
-            let v = a.unescape_and_decode_value(&self.reader)?;
-            if k.matches(&rdf::resource) {
-                object.push(self.scope().expand_iri(&v)?);
-            } else if k.matches(&rdf::nodeID) {
-                object.push(self.rename_bnode(&v)?);
-            } else if k.matches(&rdf::ID) {
-                let id = self.scope().expand_id(&v)?;
-                reification = Some(self.check_unique_id(id)?);
-            } else if k.matches(&rdf::parseType) {
-                match a.value.as_ref() {
-                    b"Resource" => parse_type = Some(&b"Resource"[..]),
-                    b"Literal" => parse_type = Some(&b"Literal"[..]),
-                    other => {
-                        let ty = String::from_utf8_lossy(other).to_string();
-                        bail!(XmlErrorKind::InvalidParseType(ty));
-                    }
-                };
-            } else if RESERVED_ATTRIBUTES_NAMES.matches(&k) {
-                bail!(XmlErrorKind::InvalidAttribute(k.value()));
-            } else {
-                attributes.insert(k, v);
-            }
-        }
-
-        // Make sure to create the right kind of object if `parseType` was
-        // explicitly given in the source document.
-        if parse_type == Some(b"Resource") && object.is_empty() {
-            object.push(self.new_bnode());
-        } else if parse_type == Some(b"Literal") {
-            if object.is_empty() && attributes.is_empty() {
-                let xmlliteral = self.factory.borrow_mut().copy(&rdf::XMLLiteral);
-                let mut scope = self.scope_mut();
-                scope.datatype = Some(xmlliteral);
-            } else {
-                bail!(XmlErrorKind::InvalidParseType("Literal".to_string()));
-            }
-        }
-
-        // Extract subject and object of the triple
-        let sbj = self.parents.last().unwrap().clone();
-        let obj = match object.len() {
-            0 if !attributes.is_empty() => self.new_bnode(),
-            1 => object.last().unwrap().clone(),
-            0 if attributes.is_empty() => self.scope().new_literal(String::new())?,
-            _ => bail!(XmlErrorKind::AmbiguousSubject),
-        };
-
-        // Add the triple and all subsequent triples as attributes
-        self.triples
-            .push_back(Ok([sbj.clone(), pred.clone(), obj.clone()]));
-        for (prop, value) in attributes.into_iter() {
-            let literal = self.scope().new_literal(value)?;
-            self.triples.push_back(Ok([obj.clone(), prop, literal]));
-        }
-
-        // Reify the triple if needed.
-        if let Some(id) = reification {
-            self.reify(id, sbj, pred, obj);
-        }
-
-        Ok(())
-    }
-
-    fn resource_empty(&mut self, e: &BytesStart) -> Result<()> {
-        self.predicate_empty(e)
-    }
-
-    fn collection_item_empty(&mut self, e: &BytesStart) -> Result<()> {
-        self.collection_start(e)?;
-        self.state.pop();
-        self.collection_item_end()
-    }
-
-    // ---
-
-    fn reify(
-        &mut self,
-        id: Term<F::TermData>,
-        sbj: Term<F::TermData>,
-        pred: Term<F::TermData>,
-        obj: Term<F::TermData>,
-    ) {
-        let mut factory = self.factory.borrow_mut();
-        let ty = factory.copy(&rdf::type_);
-        let subject = factory.copy(&rdf::subject);
-        let predicate = factory.copy(&rdf::predicate);
-        let object = factory.copy(&rdf::object);
-        let stmt = factory.copy(&rdf::Statement);
-
-        self.triples.push_back(Ok([id.clone(), ty, stmt]));
-        self.triples.push_back(Ok([id.clone(), subject, sbj]));
-        self.triples.push_back(Ok([id.clone(), predicate, pred]));
-        self.triples.push_back(Ok([id, object, obj]));
-    }
-}
-
 impl<B, F> Iterator for RdfXmlSource<B, F>
 where
     B: BufRead,
@@ -1273,8 +244,9 @@ where
                 Ok(Event::Text(t)) => self.handler.element_text(&t),
                 Ok(_) => (),
                 Err(e) => {
-                    let kind = self::error::ErrorKind::XmlError(e);
-                    self.handler.triples.push_back(Err(Error::from(kind)));
+                    self.handler
+                        .triples
+                        .push_back(Err(e.at_location(&self.handler.reader)));
                 }
             }
         }
